admin delete alerts | aprrove button visibility

+++ Server/Admin/src/views/Centers/servicecen.js
@@ -314,12 +314,18 @@ class garage extends Component {
       });
     };
 
-    // Delete User
-    const deleteUser = (id) => {
+  // Delete User
+  const deleteUser = (id) => {
+    if (window.confirm("Are you sure you want to delete this record?")) {
+      // Save it!
       const db = firebase.firestore();
       db.collection("Services").doc(id).delete();
       window.setTimeout(redirect, 2000);
-    };
+    } else {
+      // Do nothing!
+      console.log("Thing was not deleted.");
+    }
+  };
 
     const redirect = () => {
       window.location.reload();
